requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
PyJWT==2.9.0
requests==2.32.3